requirements.txt
asgiref==3.8.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
drf-spectacular==0.27.0
coverage==7.6.5
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
